+++ src/MicroElements.Metadata/Metadata/Parsing/CollectionParser.cs
@@ -3,14 +3,18 @@
 
 using System;
 using System.Linq;
+using MicroElements.Metadata.Serialization;
 
 namespace MicroElements.Metadata.Parsing
 {
     public class CollectionParser : IValueParser
     {
-        public CollectionParser(Type type)
+        private readonly IParserRuleProvider _parserRuleProvider;
+
+        public CollectionParser(Type type, IParserRuleProvider parserRuleProvider)
         {
             Type = type;
+            _parserRuleProvider = parserRuleProvider;
         }
 
         /// <inheritdoc />
@@ -19,6 +23,9 @@ namespace MicroElements.Metadata.Parsing
         /// <inheritdoc />
         public IParseResult ParseUntyped(string? source)
         {
+            if (source is null)
+                return null;
+
             string[] strings = source.TrimStart('[').TrimEnd(']').Split(new[] { ',', ' ' }, StringSplitOptions.RemoveEmptyEntries);
 
             Type elementType = typeof(string);
@@ -28,7 +35,8 @@ namespace MicroElements.Metadata.Parsing
             if (elementType == typeof(string))
                 return ParseResult.Success(strings);
 
-            IValueParser parser = GetParser(elementType);
+            IParserRule? parserRule = _parserRuleProvider.GetParserRule(elementType);
+            IValueParser parser = parserRule.Parser;
 
             var elements = strings
                 .Select(s => parser.ParseUntyped(s))
@@ -36,18 +44,15 @@ namespace MicroElements.Metadata.Parsing
 
             if (elements.All(result => result.IsSuccess))
             {
-                var values = elements.Select(result => result.ValueUntyped).ToArray();
-                return ParseResult.Success<object>(values);
+                var values = elements.
+                    Select(result => result.ValueUntyped)
+                    .ToArrayOfType(elementType);
+                return new ParserResultUntyped(values.GetType(), isSuccess: true, valueUntyped: values, error: null);
             }
 
             var errors = elements.Where(result => result.Error != null).Select(result => result.Error).ToArray();
 
             return new ParseResult<object>(isSuccess: false, value: null, error: errors[0]);
-        }
-
-        private IValueParser GetParser(Type elementType)
-        {
-            throw new NotImplementedException();
         }
     }
 }