+++ src/MicroElements.Metadata/Validation/IValidationRule.cs
@@ -72,7 +72,7 @@ namespace MicroElements.Validation
         public static string GetTypeName(this Type type)
         {
             return GetFriendlyName(type);
-            return DefaultMapperSettings.Instance.GetTypeName(type);
+            return DefaultMetadataSerializer.Instance.GetTypeName(type);
         }
 
         public static string GetFriendlyName(this Type type)